refactor(panier): modifie pour panier
what:
Modifie le panier pour ajouter le total de la commande.
Modifie la page choixArticles pour afficher le total et le tableau.

+++ projet-achat-web/src/main/java/model/LigneArticle.java
@@ -4,7 +4,7 @@ public class LigneArticle {
 
 	private int quantite;
 	private Article article;
-	private double somme;
+	private int somme;
 
 	public LigneArticle(int quantite, Article article) {
 		this.quantite = quantite;
@@ -12,12 +12,12 @@ public class LigneArticle {
 		sommePrix();
 	}
 
-	private double sommePrix() {
+	private int sommePrix() {
 		somme = quantite * article.getTarif();
 		return somme;
 	}
 
-	public double getSomme() {
+	public int getSomme() {
 		return somme;
 	}
 
@@ -37,7 +37,7 @@ public class LigneArticle {
 		this.article = article;
 	}
 
-	public void setSomme(double somme) {
+	public void setSomme(int somme) {
 		this.somme = somme;
 	}
